Split shared preferences helper class
